cast

+++ src/Space_Partitioning.cpp
@@ -4,5 +4,5 @@
 
 void createGrid(int screen_width, int screen_height)
 {
-    sf::Vector2f screen_middle = {screen_width / 2, screen_height / 2};
+    sf::Vector2f screen_middle = {static_cast<float>(screen_width) / 2, static_cast<float>(screen_height) / 2};
 }
